Pequena refatoração
